backend/pyproject.toml
[project]
name = "residential-energy-intelligence-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"


dependencies = [
    "torch==2.8.0",
    "torchvision==0.23.0",
    "torchaudio==2.8.0",
    "pandas>=2.3.3",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "numpy>=2.3.5",
    "python-docx>=1.2.0",
    "robyn>=0.72.2",
    "scikit-learn>=1.8.0",
    "pyecharts>=2.0.9",
    "openpyxl>=3.1.5",
    "weasyprint>=67.0",
    "langchain>=1.2.0",
    "openai>=2.14.0",
    "anthropic>=0.75.0",
    "ollama>=0.6.1",
    "apscheduler>=3.11.2",
    "h5py>=3.15.1",
    "hdf5plugin>=6.0.0",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
    "pathlib>=1.0.1",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true


